[codec] Avoid creating base class instances ever
This fixes one cause of fxbug.dev/67247.  The H264Decoder
class we're using for DPB management wasn't using
CreateH264Picture() in one path, but we were assuming we
would always have an AmlogicH264Picture, so we ended up
accessing an H264Picture as if it were an
AmlogicH264Picture, with predictably unpredictable
results.

BUG: fxbug.dev/67247
Change-Id: If7aee2366f58123ac55a324673aea1981ddd9441
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498218
Reviewed-by: Austin Foxley <afoxley@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>