[libtriage] Support Not(Missing(...)) for unknown types.
Some values (like histograms) cannot currently be handled by libtriage,
however, we want to be able to assert that they are at least present.

This change adds a new MetricValue::NotMissing alternative that may be
used for types that are present but not understood.

Change-Id: Ieb565110d4e8b5ad9c2b4ba39b4d952985a58677
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490865
Commit-Queue: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>