[wlan] Immediately identify an AP as Open if the privacy bit is not set
This change removes the possibility that an unusual network
configuration could be identified as WPA2 when the privacy bit is not
set. This privacy bit itself conventionally determines if a network is
Open or not.

The mocked Wpa2NoPrivacy security type is removed and no longer
supported. This security type is a corner case in IEEE 802.11-2016
that we do not test support for, and it should not be supported since
it implies a WPA2 connection that is not encrypted.

Test: Existing test coverage is sufficient.

Change-Id: I28f45d5fc3fe4bd3713ea5eed4e896efdc65adae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470421
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>