[fidlc] Conditionally remove support for protocols as types
This CL modifies fidlc so that under the new syntax, protocols
cannot be interpreted as types. To do this, it refactors
CompileTypeConstructor to validate the "sort" (see
ExpectedSort in flat_ast.h) of the thing it is compiling.

In the old syntax, CompileTypeConstructor allows either types
or protocols, whereas in the new syntax it only allows types.
In either case, we never allow using services as a type. Also,
`CompileService` leverages this to validate that its members
are protocols.

Test: fidldev test fidlc
Change-Id: Ibc1a0ec9a263b9633c122e817a52a8fb00edad24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486197
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>