[build][tracer] Unwrap known wrapper commands
The action-tracer ignore logic uses the name of the underlying script,
but we are not interested in command-wrappers that may appear before it.
This peels apart wrapper commands to get to the tool of interest.

This is needed to make the ignore logic works in the presence of command
wrappers, e.g. output_cacher.py.

Test: (enable tracing in fx set; fx build)
Test: fx test build/tracer

Change-Id: I24d7f79ba281170f92b1fd53d95e7cec51fc2b02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497610
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: David Turner <digit@google.com>