[runner] Don't set a pgid when forwarding os.Stdin to subprocess
I don't fully understand why, but unless I do this, subprocesses are not
able to receive stdin piped from the parent processes, even when
`cmd.Stdin` is set to `os.Stdin`.

This is a problem for interactive goma login when triggered by a
subprocess of `fx-set`.

This is effectively a combined revert+reland of fxrev.dev/498038, which
attempted to accomplish the same thing using a different approach.
Conditionally setting `Noctty` solved the problem on Linux but not on
Mac.

I've tested this approach on both Linux and Mac and it seems to work.

Bug: 68465
Change-Id: I41c75abf945648e5d4392e0e0791f1de6709c6b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498141
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>