[kernel][crashlog] Fix uptime units in the crashlog startup banner.
Fix the uptime units printed out in the crashlog startup banner.
Before this, the banner reported uptime in nanoseconds, but the
message said that the units were "ms", which is incorrect.  Now, we
report the values in seconds, but with 3 digits of msec resolution as
well.  This is the format used for presenting timestamps in kernel log
messages, so hopefully it will be easy for folks to read and reason
about.

Change-Id: I0d675f2ed0e99cd4cb5734dfbb9824022f5e3651
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497649
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>