[debugger] Show thread context on global pause.
Outputs the current thread context on a global pause. This wasn't
getting output in the common case of pausing immediately after attaching
because there was no current thread (the threads had not be synced from
the debug_agent before the pause was issued).

This change always outputs the current thread after the command
completes which avoids the issue, and also is more consistent if the
current thread changes during the time the pause was being issued.

Allow "global" annotations for "pause" and "continue" commands. This is
technically redundant but I found myself wanting to be explicit and was
surprised when this issued an error.

Change-Id: Iada310e1eda39e4f06155c06dd6b1b202189d98b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499060
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>