[vulkan][gtest] Remove usages of gtest internal methods
Removes some usages of gtest internal methods found when
attempting to roll in an updated version of gtest.
The internal methods have been updated in gtest and the
macro in vulkan breaks with the roll.

Testing: Verify test succeeds with current and new
version of gtest.

Change-Id: I91b2b1a6499f0117046593328437dc2b5d194093
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498708
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>