[kernel][boot-shim][qemu] set the IPI base to 0 instead of 12
ARM GIC defines interrupt 0-15 as SGI (Software Generated Interrupts),
which are effectively used as inter-processor interrupts (IPIs).
Technically they're all free to use, but for historical reasons that
are largely untrue in retrospect, early code had assumed that lower IPI
numbers are less desirable. We have been setting the IPI 'base' (number
at which we start allocating IPIs) in the high range of 0-15 (12 in
this case) as a result.

Upon re-reading the GIC manuals it says quite specifically that you
should generally be allocating near the bottom, and more specifically
0-7 should be used for non secure IPIs and 8-15 for secure. Since
we're running in a non-secure EL1 we should favor towards the bottom.

Tweak the qemu IPI base to 0, but more importantly add a printf to
print the IPI base so we can start trying to assess where we've been
allocating up until now more easily for future tweaks towards 0 on
other platforms. A quick glance through a few other platforms we seem
to generally be already allocating near 0 for most.

No functional change, just switches to IPI 0,1,2,3 instead of
12,13,14,15. Nothing in QEMU cares about this so it works just fine.

Change-Id: Iabcddeb1034f0be255cf11d8d97a60790dd3d640
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497608
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>