[rfc] Minor edits for clarity in RFC-0072.
The term "base package" is overloaded, in that it could mean "the
package called `base`" or "a package in the base package set". I got
confused in one circumstance, so I made it explicit.

I also found a use of `system_image`, which I believe is called `base`
in all other circumstances.

Change-Id: If5fd7987c3fe281105efc0a4da059fedc6e0ccf3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499318
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>