[zbi] Support packing STORAGE_KERNEL items
Like with STORAGE_BOOTFS items, STORAGE_KERNEL items are stored in
the BOOTFS format. We introduce support to pack the former using the
same command-line ergonomics as the latter. We do this in part by
introducing a new `--filestype` flag to allow for the selection of one
of the BOOTFS types.

Bug: 68585
Change-Id: Idad540e72aa547a8325e12fa3b92550ce8e5aaaa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498698
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>