[time] Reset time source puppet state when timekeeper restarts it
http://fxrev.dev/493010 updates Timekeeper to emit
error logs when a time source emits a stale sample.
In our tests where timekeeper restarts a time source,
the puppet source retains the old sample from before
the restart and sends it as soon as timekeeper
reconnects.
This updates the puppet to reset its state when it
receives a new connection.

Change-Id: I92b4e20476486a8e2cc88c60eb4f61357ec8cb7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499390
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>