[kernel] Fix mexec case when RAMDISK needs to be relocated
The mexec code moves the RAMDISK image to a different address if
forced by a kernel command line option or if required to avoid
overlap with the kernel image.  In the latter case, it was
setting up to do the copy and then acting like it had done the
copy but failing to actually do the copy.

Bug: 32414
Change-Id: Ib59ba29a7151b19ec801fca2b44288a4812d8040
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499600
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>