[build] action wrapper check for reproducibility
Wrapping an action with this check is also action-traceable,
hence this transformation can be applied first.

The enable checking of all actions:
	fx args: check_repeatability = true

To mark a suppression in an invocation:
	repeatable = false

Initially, the output_cacher.py script is --disabled for the sake
of making sure the GN is correct.

action_tracer.py: Ignore all file accesses to the --temp-dir.

Test (4 combinations):
  fx args
    (edit)
    build_should_trace_actions={true,false}
    check_repeatability={true,false}
  fx build

Change-Id: If53942246dea91d305b014b664c7e5d4ba1410f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497618
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>