[kernel][wdt] Fix a hypothetical issue with a suppressed WDT
Even if WDT petting is suppressed, make sure to schedule the next pet
timer relative to just before when we _would_ have pet the dog.  We
want the last_pet_time to accurately reflect when we last actually pet
the dog, but if we schedule our next timer based on last_pet_time, and
petting is suppressed, this could cause us to end up continuously
scheduling a timer in the past, effectively lockup up a core.

Right now, we only ever suppress dog petting during the generation of
a crashlog during a SW WATCHDOG event.  This is our failsafe to ensure
that the system _will_ reboot, even if the crashlog generation hangs.
Even so, it would be best to not end up simply spinning another core
in timer IRQ processing while this happens.

Change-Id: I58ab695d075470c890998c0f163d0cf9206b9802
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498459
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>