[settings] Expose service messagehub to env creator
In order to watch the service MessageHub via a broker,
the creator of the environment needs to have access to the message
hub. This CL changes the return so that it returns the message hub
along with the nested environment.

Fixed: 71062
Test: fx test -o setui_service_tests
Multiply: setui_service_tests

Test added to check that the message hub returned from environment
creation is the same instance as the one in the rest of the service.
Uses a test agent to receive a broadcasted message over the message hub
returned by environment creation and verifies that the test agent can
receive it.

Change-Id: I4dc20f26b6f552d9a06178ce29e8f3a111b0dcc0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492019
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Sijie Chen <sijiec@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: William Xiao <wxyz@google.com>