[kernel][lockup_detector] Swap the order of the condition checks.
Swap the order of the condition checks, preferring to check the
critical section condition before the heartbeat condition.  The
thinking here is that the CS check contains more context about where
things went wrong than the heartbeat condition does.  If we are going
to generate a crashlog and reboot, and both conditions have been
violated in a fatal fashion, we would rather report the crashlog
condition than the heartbeat condition.

Change-Id: I5c4806b6b671c595675b58edfd522e48d491ce32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499080
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>