[timekeeper] Increase some warnings to errors.
Previously we'd tended to use warnings for conditions that really
shouldn't be occuring but were not our fault and which we could recover
from.

However, RFC-0003 states "Note that the bug might not be in the code
issuing the log statement. The bug may be in one of the callers or one
of its dependencies." Given that and given that errors now have quite a
bit more tooling to help track and isolate them I'm bumping a few
conditions up from warning to error.

Change-Id: Ifd641d4c101bcce1f7b120cd4ba39f7e36c0f9e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493010
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>