[build][repro] Run, backup, run, compare
Previously, we renamed declared outputs in the re-run of the command.
However, it turns out that some actions outputs' contents are sensitive
to output-path/name differences on the command line.
So, we leave the command-line alone, and use the renamed location
to make a backup copy of declared outputs before rerunning.

Change-Id: I112c1256a3c6941bac4913f5f97363eafecb3eed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498352
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>