[zbi] Do not rely on item type to determine BOOTFS-ness
Today we check whether the type is ZBI_TYPE_STORAGE_BOOTFS in
determining whether to display an item as having a BOOTFS payload. We
will soon invalidate the assumption that BOOTFS-ness can be determined
solely by looking at type* and it also sensible for `zbi` to not make
assertions on BOOTFS-ness that extend beyond the ZBI spec (least of all
in these diagnostic modes).

*We will introduce support to pack ZBI_TYPE_STORAGE_KERNEL items in the
BOOTFS format, but it is not necessarily true that all such items will
be in that format.

Bug: 68585
Change-Id: I031ef34feb4452b95eba14dee5f47c2cf5c20609
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499889
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>