[gigaboot] add libavb to build
Adds libavb to the device-side build and ports the necessary sysdeps.

libavb was designed to be forked and modified so this is a little
awkward, but better for our purposes so we don't have to create a second
copy of the code in-tree. Luckily we already have a lot of Posix APIs
implemented so this just needed to add a few more.

In the longer-term, we'll depend on the full zircon_boot library instead
of libavb alone, but this unblocks channel-in-vbmeta work until we're
ready for that.

Bug: 71571
Change-Id: I3c0eed2392ad6091bc857e086e6dda400bbafd36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498358
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>
Commit-Queue: David Pursell <dpursell@google.com>