[zbi] Support packing STORAGE_KERNEL items
Like with STORAGE_BOOTFS items, STORAGE_KERNEL items are stored in
the BOOTFS format. We introduce support to pack the former using the
same command-line ergonomics as the latter. We do this in part by
introducing a new `--filestype` flag to allow for the selection of one
of the BOOTFS types.

This is an reland of
https://fuchsia-review.googlesource.com/c/fuchsia/+/498698, which had a
buggy assumption in the display logic. That logic has since been
refactored in a parent change and this change differs from the original
only in no longer touching that logic.

Bug: 68585
Change-Id: Iba35d9192b7715f1c201317b2d8faa7d41807e63
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499890
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>