[vmm] Hide scanout image in guest view if scanout is disabled.
If virtio_gpu driver sends SetScanout command with resource_id
of 0, this means the provided scanout should be disabled and
scanout_source_height_ and scanout_source_width_ will be reset
to 0.

In that case we shouldn't display the scanout buffer, instead we
could just set the background to a plain color (black).

TEST: on workstation.x64: run terminal; run debian_guest; close terminal
      No crash on guest.
Change-Id: I0b0a7363a5ec82c7248130765d9484e1aa6ea3bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499880
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Yilong Li <liyl@google.com>