[vulkan][gtest] Add back Vulkan test macro
This macro was originally removed as it uses gtest internal methods
that prevented a roll. Now that the roll is complete, this adds back
the macro, updated for the new methods. Note that this still relies on
gtest-internal methods.

This change requires the new version of gtest, rolled in tqr/340050.

Change-Id: I9c8de5e8930533e7871089e53fc2b92f2e4b48cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499387
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>