[logs] Root archivist logs to debuglog to avoid losing its own logs.
When the root Archivist crashes in the field we currently don't get
any logs from it for the current boot log. If there's a subsequent
crash we'll get them in the previous boot logs if we're lucky, but it's
much safer for the logging service to send its own error logs straight
to the kernel.

Fixed: 69559
Change-Id: I5f6866425bfa9b9672675e8a83e86c991de55af3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499886
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>