[system-updater] enforce the OTA backstop
In this change, we put it all together and enforce the OTA backstop.

Test: fx test -o system-updater-tests
Test: fx test -o system-updater-integration-tests
Test: manual test by OTA to epoch 1, and verify downgrade OTA to epoch 0
will fail with unsupported downgrade. Note this will not print out with
`fx ota` because it's not exposed in the fuchsia.update state...but it
prints in the logs. fxbug.dev/72118 proposes DX improvements for this.

Bug: 68408

Change-Id: Ied19cc437e7d9d8eb8426d1a1d83f7b198ae8d9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499981
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>