[gtest] Reenable test skipped verification
Previously, this test was disabled to enable a gtest roll as the roll
introduced a new log that broke the test. This reenables the test,
updated to account for the new log.

This change requires the new version of gtest, rolled in tqr/340050

Testing: fx test gtest-runner-integration-test test_manager_test

Change-Id: I2512f1d656b8bde2fc2564bb4bd63a7bf142a779
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499384
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>