[sl4f] Split trace processing test data out into json files
The large dart file was impossible to view in code search and code
review, and since many of the traces were a single line they were liable
to cause problem for grep tools, with no particularly easy way to
exclude the file.

This change splits the traces out to pretty-printed json files, which
should allow tools to process diffs, and since they are json files they
can be excluded e.g. with ripgrep's -T option.

Cpu and Gpu metrics needed to be modified to have an explicit conversion
to double when reading argument values, to fix 'int is not a subtype
of double' errors.

Fixes: 71931
TEST: fx test sl4f_client_test

Change-Id: I5fd0018dcd9c448795be5631b09c8bce07770bfa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500078
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>