[time][test] Remove wait for UTC
Removes a previously added hack to wait for UTC before running a test.
gtest used to use UTC time to measure test duration. This could result
in inaccurate measures if the clock was adjusted during a test run. Now
that gtest uses monotonic time, this hack is no longer necessary.

Note that this increases the chance that a test will observe UTC time
jumping or not moving at all, and may uncover flakes in tests that
incorrectly assume UTC time is monotonic.

This change requires the new version of gtest, rolled in tqr/340050.

Fixed: 65170
Bug: 66988

Change-Id: Iee26b68552a1a2fdf816e47b1eff4dfda59ada70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498345
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>