[checklicenses] Match file list fix.
Match.File should contain a list of files that that specific Match instance matches.

If it's a project-level license, we don't want to also include all of the project files in that list, otherwise it becomes difficult to verify correctness later. I've created a second list in the Match struct to hold those files (though tbh right now it's not being used for anything, so we may be able to get rid of it entirely)

Change-Id: I4377fb42e0c22bb96b6ce532c76b459dce36274a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500799
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>