[tefmocheck] split swarming output into per-test files
This is a prerequisite for associating check failures with specific
tests. See attached bug for details.

And add those files to the test's OutputFiles in the summary. This will
expose the files to ResultDB, Sponge, and the recipes, which may be
useful.

Bug: 67640
Change-Id: I3be61bf99b75a5c4747d91e721ba9a2b10320a08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499558
Commit-Queue: Gary Miguel <garymm@google.com>
Fuchsia-Auto-Submit: Gary Miguel <garymm@google.com>
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>