[fdevtools] fix script when fdt has already been extracted
This fixes https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=71284

The issue is that we were checking to see if we had already extracted something from a zip file, if we had then we were just executing the binary that was already extracted. This is wrong because it means even when we download something new we will execute the old thing because it has already been extracted. This instead deletes the extracted directory and always unzips to that location. This is slightly worse on subsequent runs when we are not downloading the zip anew, but is always safe and is not really noticeable. This is also the exact same thing we do for "fx fdt".

I also had to fix the gn-bash-test-lib.sh because the definition of realpath doesn't jive with the usage by btf because it calls realpath with a -m as the first argument. I presume that was a later change that happened before this overwrite was written.

Change-Id: Ie6aebcf59cabbbeaff5570897e684de5e46da765
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496444
Commit-Queue: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Jayson Adams <shrike@google.com>
Reviewed-by: Omer Lev-ran <omerlevran@google.com>