[build] Remove variant_target parameter from toolchain_utils_action().
This parameter is obsolete and completely ignored. It's a memory of
the previous ZN build, which didn't implement variant builds in the
same way as the Fuchsia build. So remove all the places where
it was previously defined. This should not change the build.

Bug: None
Change-Id: I0e77079e3cdf08892a1334e8ea540b92e656a8a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499938
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>