[hoist] add facility to disable autoconnect
The ffx frontend is in control of the ascendd host, and wants to
restart it on upgrade. In order to assert when the node goes away
(promptly), the frontend needs awareness of the unix socket link,
thus it wants to take control. This API enables it to do so.

Change-Id: I275bb5151eabde9968c0e00321c032730c2e65ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496448
Commit-Queue: James Tucker <raggi@google.com>
Reviewed-by: Craig Tiller <ctiller@google.com>