[audio_core][tests] Show all results, update baseline
Update (tighten up) the saved fidelity baseline thresholds to
reflect the state of the current system.

By default, fidelity results are not displayed at all unless the
full-spectrum of frequencies is tested, so if this is NOT the
case, but '--recap' IS specified, it is reasonable to display
all results (not just a subset) -- but only for the 4 summary
frequencies, not all 52.

Test: fx test -o audio_fidelity_tests -- --recap

Change-Id: I67021a66f238079c633235e8d5d2ea5ac72d9a64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499887
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>