[tefmocheck] Handles when no summary.json exist
The recipe skips the -test-summary-json flag when summary.json doesn't
exist on disk. This caused a panic in case of test failure without a
summary.json. Silently skip in this case, as the test had failed anyway.

It would be preferable to test this part but this requires a bit of
refactoring to create the equivalent of a mainImpl(). This can be done
as a follow up.

Bug: 72235
Change-Id: If3677ab9ab7dbc9ce043624d11cb6428fd69cadf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501718
Fuchsia-Auto-Submit: Marc-Antoine Ruel <maruel@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Marc-Antoine Ruel <maruel@google.com>