[smc] Add SMC return values to overrun warning
The contents of the x0-x3 after an SMC instruction will contain the
return parameters of the call. For yielding calls, the values will
commonly contain the reason for why the SMC returned. This could be that
the yielding call has completed, or it could indicate that we're
returning to process an interrupt. This CL adds the x0-x3 values to the
debug log statement that occurs when we detect the SMC call overrunning
the preemption timer.

Bug: 69037
Change-Id: I170250d1fc9604d83bf30d252d56464be7d73ede
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500961
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>