[fidlc] Enable syntax version declaration.
In cases where the --allow_new_syntax fidlc flag is set, the parser uses
the "deprecated_syntax" flag at the top of the file to determine which
syntax to use.  If the flag is not set, the parser defaults to using the
old syntax, and errors whenever a "deprecated_syntax" token is detected.

Bug: 65978
Change-Id: I51c6d8f54a90c455d2f55b04728076c7aa0e338e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492578
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>