[scenic] Use dedicated allocation for BufferCollection images
In practice, these allocations are already effectively dedicated.
Making that explicit can help drivers and may be required in the future.

This also fixes a bug (now caught by the validation layers) where the
VMO size is being used to create the memory instead of the size from the
memory requirements. On some drivers using the VMO size is incorrect.

Bug: 72081
Change-Id: Ide4b49157c8747df65044b5ba4a0a1c8fed15e1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500280
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>