[pci] Add PciAckInterrupt protocol method
This new protocol method allows a device driver to re-enable their
device's legacy interrupt before waiting. This is done so that the Bus
driver can disable a device's interrupt generation while an interrupt is
being serviced to prevent the IRQ worker being woken erronously when a
driver fails to service a request or does not exist for a given device.

This method is only necessary with userspace PCI, but a noop shim has
been added to the kernel PCI driver for migrational purposes.

Bug: 32978, 71188
Change-Id: I161a68f7db52b70d28700e2a69ee8f5a6b6a55f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472847
Commit-Queue: Christopher Anderson <cja@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>