[debugger] Work around Clang structured bindings symbol bug.
Clang structured bindings symbols are incorrect as described in the new
comment and in
  https://bugs.llvm.org/show_bug.cgi?id=49565

This workaround shows all "artificial" variables in the "locals" command
to work around this bug. The "artificial" DWARF tag is mostly just used
for "this" parameters which was already explicitly ignored by this
function, so there is currently no known down side to always showing
these.

This hides unnamed variables in "locals".

Change-Id: I985283f01d07041082dbac1b30e658a3106c7685
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500813
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>