[fx-tools] merge package_config files on demand
This CL adds an fx tool to merge 2 package_config.json files.
This tool is then used from the various dart based fx tools to
merge their package_config.json file with the one that is generated
when we roll 3p packages. This is required because the package_config.json
file specifies the dart version that the package should be compiled
against. When we roll in dart we update our 3p packages which can cause
them to get out of sync with the fx tools and break the in tree tooling.

When we invoke these tools we will create a merged package_config.json
file and use that as the basis of our dart invocation. This has the
drawback of having to execute a python script each time you run something
like fx status. We can optimize this later to run the script on
rolls and to put the outputs in the third-party dart repo.

Change-Id: Id26c01c0438bc93dae44254291c9a9a4da9b7885
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501822
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>