[bootsvc] Use valid handle for `zx_debuglog_create`.
The `zx_debuglog_create` is a (or, _should be_) a privileged syscall,
allowing access to the kernel's debug log. Long term, the goal is that
only processes with the root resource (or a similarly debuglog-specific
resource) should be able to create a debuglog object.

As a backwards compatibility concession, `zx_debuglog_create` currently
allows an invalid handle to be passed into `zx_debuglog_create`.

To make progress towards the goal of removing this concession, this
CL updates bootsvc to pass the root resource to the `zx_debuglog_create`
syscall instead of an invalid handle.

Bug: 32803
Change-Id: Ibdb2bffb3b1805f9eae28aa44260117a76dee139
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501441
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>