[ffx] improve edge condition behavior on upgrade
The upgrade path detects a new binary version by the hash of the
binary itself. The code now has explicit handling for the underlying
unix socket as a rendezvous so as to not prematurely connect to the
old daemon leading ot failures, and to improve error reporting if a
daemon exits preamturely. A race condition exists where once the
frontend promptly starts to rediscover the daemon it may still
discover the old daemon on the Overnet mesh, as the node still exists
in the peer table even though there are presently no active known
routes to it. The old daemon nodeid is now explicitly ignored in the
post-upgrade daemon discovery.

Fixed: 70486
Fixed: 57264
Change-Id: I000ef41250e1ea55f612f0125fa6f252251e7235
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496968
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Matthew Boetger <boetger@google.com>