[tefmocheck] rename per-test swarming output files
ResultDB shows the file base name in the UI, so we want to keep the
base name the same as the unified swarming output file so that it's
clear what it contains.

Also we don't need both the test name and the test index in the
directory hierarchy.

Also simplify a test by removing a loop since there was only one test
struct in the slice being looped over.

Bug: 67640
Change-Id: I46e03e0765f0bc98a7857dc393f80a842ee12873
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502053
Commit-Queue: Gary Miguel <garymm@google.com>
Fuchsia-Auto-Submit: Gary Miguel <garymm@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>