[libdoc] accept backslahes in strings
The lexer was interpreting backslahes in strings. It would have been
correct for a computer language but it is not correct for documentation.

This CL removes the code which used to interpret backslashes and add
tests to check that we can have a backslash at the end of a string.

Change-Id: I565439fabb8c86bcb0c4d25e8299aff7ec03980d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499578
Commit-Queue: Vincent Belliard <vbelliard@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>