[stress_tests] Implement limits for component manager stress tests
Component manager stress tests were biased towards component creation
and did not have limits on the number of components that could be
created. This led to the test failing because of a high handle count.

This CL changes how this test works. Components will keep being added
until a certain threshold is hit (2000 components) after which a
random component (and its subtree) will be deleted to bring the
number of components back under the threshold. This process happens
continuously.

In practice, a threshold of 2000 components means that component_manager
holds onto ~10000 handles concurrently.

Test: fx test component-manager-stress-tests-alt
Change-Id: I21d535e2a5ad451db88e36311d4b5b528098cc1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501786
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Gary Bressler <geb@google.com>