[fx test] Make fs test tolerate non-utf8 encoding
Changes the code that reads test output to make it tolerant of
non-utf8 encoded text. Instead of failing, it prints the non-
utf8 characters with diamonds: like �����

Also, adds exception handling so that errors in test runs are
caught and reported instead of exiting with a stack-trace.

Tested:
• Added an invalid UTF8 string to a test output. Eg: Added
    fmt.Print(string([]byte{147}))
  to a test in //src/lib/component/serve_test.go
• Confirmed the utf8 exception is thrown and `fx test` fails.
• Then confirmed that with the changes in this CL, the
  invalid characters do not cause an exception and the test
  passes as expected.

Fixed: 71126,69029
Change-Id: Ie16862fd1c2b7fa709021ce0f4e46caf509d77d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494324
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>
Commit-Queue: Gary Boone <gboone@google.com>