[flutter] Use fake display for tests
The test was hardcoding 200x200 for the child view and overlay logical
sizes.  We use a default logical->physical scale of 1.7, so the overlay
gets cut off if the display is < 680px in x or y, failing the test.

The solution was to use FractionallySizedBox() to stop hardcoding the
child view and overlay sizes.

BUG: 70433
BUG: 67316
TEST: Deflaking flutter-embedder-tests

Change-Id: Ib29e3d1b5b4d611b4b719f52abfa40089ae8a972
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502231
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: David Worsham <dworsham@google.com>