[gigaboot] serial input support
Gigaboot has some simple menus that can be accessed during boot, but
while it would print output to both the display and serial, it only
listened for input on an attached keyboard.

This CL adjusts the input functionality to look for either the attached
keyboard or serial, which makes it possible to interact with Gigaboot
using just a serial line, no keyboard needed.

While we're here, also slight cleanup to the boot menu formatting to
make it a little easier to read - previously all the options were on a
single line, which often got cut off over serial e.g. using Minicom.

Bug: 70465
Test: NUC boot menu works over serial and USB keyboard
Change-Id: I1aa85098f193701237da321fb3427dd800c17c9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502501
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>